==> Part_2_cian_parser_2 (data from hrefs).ipynb <==
import requests 
import re
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup

pd.set_option('display.max_columns', None)

df=pd.read_csv('hrefs_from_cian.csv',header=None)

df=df[0]

df=list(df)

df[0]

#Функция по парсингу необходимой информации по квартирам:
def get_page_data(files):
    lst = list()
    for file in files:
        data = dict()
        response = requests.get(file)
        html = response.content
        soup = BeautifulSoup(html,'html.parser')
        try:
            data['Название'] = soup.find('h1', class_=re.compile('\.*--title--\.*')).text
        except:
            data['Название'] = np.nan
        try:
            data['Описание'] = soup.find('p', class_=re.compile('\.*--description-text--\.*')).text
        except:
            data['Описание'] = np.nan
        try:
            data['Адрес'] = soup.find('address', class_=re.compile('\.*--address--\.*')).text.replace('На карте', '')
        except:
            data['Адрес'] = np.nan
        try:
            data['Метро'] = soup.find('a', class_=re.compile('\.*--underground_link--\.*')).text
        except:
            data['Метро'] = np.nan
        try:   
            data['Время до метро'] = soup.find('span', class_=re.compile('\.*--underground_time--\.*')).text
        except:
            data['Время до метро'] = np.nan
        try:
            data['Цена'] = soup.find('span', class_=re.compile('\.*--price_value--\.*')).text[:-1].replace('\xa0', '')
        except:
            data['Цена'] = np.nan
        try:
            data['Цена за квадрат'] = soup.find('div', class_=re.compile('\.*--price_per_meter--\.*')).text[:-5].replace('\xa0', '')
        except:
            data['Цена за квадрат'] = np.nan
        try:    
            info = soup.find_all('div', class_=re.compile('\.*--info--\.*'))
            for div in info:
                value = div.find('div', class_=re.compile('\.*--info-value--\.*')).text
                name = div.find('div', class_=re.compile('\.*--info-title--\.*')).text
                data[name] = value
        except:
            pass
        try:    
            item = soup.find_all('li', class_=re.compile('\.*--item--\.*'))
            for span in item:
                value1 = span.find('span', class_=re.compile('\.*--value--\.*')).text
                name1 = span.find('span', class_=re.compile('\.*--name--\.*')).text
                data[name1] = value1
        except:
            pass
        try:    
            item1 = soup.find_all('div', class_=re.compile('\.*--item--\.*'))
            for div1 in item1:
                value2 = div1.find('div', class_=re.compile('\.*--value--\.*')).text
                name2 = div1.find('div', class_=re.compile('\.*--name--\.*')).text
                data[name2] = value2
        except:
            pass
        lst.append(data)
    df = pd.DataFrame(lst)
    df.to_csv('cian_data.csv', index=False)
    return df

#### Проба 

def get_page_data1(files):
    lst = list()
    data = dict()
    response = requests.get(files)
    html = response.content
    soup = BeautifulSoup(html,'html.parser')
    try:
        data['Название'] = soup.find('h1', class_=re.compile('\.*--title--\.*')).text
    except:
        data['Название'] = np.nan
    try:
        data['Описание'] = soup.find('p', class_=re.compile('\.*--description-text--\.*')).text
    except:
        data['Описание'] = np.nan
    try:
        data['Адрес'] = soup.find('address', class_=re.compile('\.*--address--\.*')).text.replace('На карте', '')
    except:
        data['Адрес'] = np.nan
    try:
        data['Метро'] = soup.find('a', class_=re.compile('\.*--underground_link--\.*')).text
    except:
        data['Метро'] = np.nan
    try:   
        data['Время до метро'] = soup.find('span', class_=re.compile('\.*--underground_time--\.*')).text
    except:
        data['Время до метро'] = np.nan
    try:
        data['Цена'] = soup.find('span', class_=re.compile('\.*--price_value--\.*')).text[:-1].replace('\xa0', '')
    except:
        data['Цена'] = np.nan
    try:
        data['Цена за квадрат'] = soup.find('div', class_=re.compile('\.*--price_per_meter--\.*')).text[:-5].replace('\xa0', '')
    except:
        data['Цена за квадрат'] = np.nan    
    try:    
        info = soup.find_all('div', class_=re.compile('\.*--info--\.*'))
        for div in info:
            value = div.find('div', class_=re.compile('\.*--info-value--\.*')).text
            name = div.find('div', class_=re.compile('\.*--info-title--\.*')).text
            data[name] = value
    except:
        pass
    try:    
        item = soup.find_all('li', class_=re.compile('\.*--item--\.*'))
        for span in item:
            value1 = span.find('span', class_=re.compile('\.*--value--\.*')).text
            name1 = span.find('span', class_=re.compile('\.*--name--\.*')).text
            data[name1] = value1
    except:
        pass
    try:    
        item1 = soup.find_all('div', class_=re.compile('\.*--item--\.*'))
        for div1 in item1:
            value2 = div1.find('div', class_=re.compile('\.*--value--\.*')).text
            name2 = div1.find('div', class_=re.compile('\.*--name--\.*')).text
            data[name2] = value2
    except:
        pass
    lst.append(data)
    df = pd.DataFrame(lst)
    #df.to_csv('cianan.csv', index=False)
    return df

get_page_data1('https://spb.cian.ru/sale/flat/237138478/')

get_page_data(['https://spb.cian.ru/sale/flat/237138478/','https://spb.cian.ru/sale/flat/235416723/'])
