==> Part_3_data_analyze.ipynb <==
import re
import pandas as pd
import numpy as np

pd.set_option('display.max_columns', None)

cian_df=pd.read_csv('cian_data.csv')

cian_df=cian_df.drop_duplicates()
cian_df.reset_index(inplace=True)
cian_df=cian_df.drop('index', axis=1)

cian_df.describe()

cian_df.info()

cian_df.columns

# Рассмотрим каждый признак, очистим данные и заполним пропущенные значения

#Удалим признаки, где слишком мало данных
cian_df = cian_df.drop(['Газоснабжение', 'Парковка', 'Мусоропровод',
       'Строительная серия', 'Срок сдачи', 'Отделка', 'Комната',
       'Всего комнат в квартире', 'Комнат в продажу', 'Площадь комнаты',
       'Размер доли','Планировка'], axis=1)

### Обработка признака "Общая"

cian_df['Общая'].isna().sum()

#Удалим пропущенные значения общей площади
cian_df=cian_df.dropna(subset=['Общая'])

cian_df['Общая'] = cian_df['Общая'].apply(lambda x: float(x.replace('\xa0м²','').replace('1 ','').replace(',','.').replace(' ','.')))

### Обработка признака "Жилая"

#Обработка пропущенных значений жилой площади:
cian_df1=cian_df.dropna(subset=['Жилая'])
cian_df1['Жилая'] = cian_df1['Жилая'].apply(lambda x: float(x.replace('\xa0м²','').replace('1 ','').replace(',','.').replace(' ','.')))
koeff=cian_df1['Общая'].mean()/cian_df1['Жилая'].mean()
cian_df['Жилая'][cian_df['Жилая'].isnull()] = cian_df['Общая'][cian_df['Жилая'].isnull()]/koeff
cian_df['Жилая'] = cian_df['Жилая'].apply(lambda x: str(x))
cian_df['Жилая'] = cian_df['Жилая'].apply(lambda x: float(x.replace('\xa0м²','').replace('1 ','').replace(',','.').replace(' ','.')))

cian_df['Жилая'].min()

#Удалим явный выброс:
cian_df['Жилая'][cian_df['Жилая']==0.0]=np.nan
cian_df = cian_df.dropna(subset=['Жилая'])

### Обработка признака "Кухня"

#Обработка пропущенных значений площади кухни:
cian_df2=cian_df.dropna(subset=['Кухня'])
cian_df2['Кухня'] = cian_df2['Кухня'].apply(lambda x: float(x.replace('\xa0м²','').replace('1 ','').replace(',','.').replace(' ','.')))
koeff1=cian_df2['Общая'].mean()/cian_df2['Кухня'].mean()
cian_df['Кухня'][cian_df['Кухня'].isnull()] = cian_df['Общая'][cian_df['Кухня'].isnull()]/koeff1
cian_df['Кухня'] = cian_df['Кухня'].apply(lambda x: str(x))
cian_df['Кухня'] = cian_df['Кухня'].apply(lambda x: float(x.replace('\xa0м²','').replace('1 ','').replace(',','.').replace(' ','.')))

cian_df['Кухня'].min()

cian_df['Кухня'].max()

### Обработка признака "Название"

cian_df['Название'] = cian_df['Название'].apply(lambda x: x.split(',')[0])

cian_df['Название'].value_counts()

cian_df['Название'][(cian_df['Название']=='Апартаменты свободной планировки')|(cian_df['Название']=='19/353 квартиры')|(cian_df['Название']=='Квартира свободной планировки')]=np.nan
cian_df=cian_df.dropna(subset=['Название'])

def rename(x):
    if x == 'Апартаменты-студия':
        a='Студия'
        return a
    if x == '1-комн. апартаменты':
        a='1-комн. квартира'
        return a
    if x == '2-комн. апартаменты':
        a='2-комн. квартира'
        return a 
    if x == '3-комн. апартаменты':
        a='3-комн. квартира'
        return a
    if x == '4-комн. апартаменты':
        a='4-комн. квартира'
        return a
    if x == '5-комн. апартаменты':
        a='5-комн. квартира'
        return a
    if x == 'Многокомнатные апартаменты':
        a='Многокомнатная квартира'
        return a
    else:
        return x

cian_df['Название'] = cian_df['Название'].apply(rename)
cian_df['Название'].value_counts()

### Обработка признака "Цена за квадрат"

def replace_price_per_square(x):
    try:
        x=x.replace(' ','')
    except:
        pass
    return x

cian_df['Цена за квадрат'] = cian_df['Цена за квадрат'].apply(replace_price_per_square)
cian_df['Цена за квадрат']=cian_df['Цена за квадрат'].apply(lambda x:int(x))

#Но вообще он очень сильно коррелирует с целевой переменной, поэтому удаляем:
cian_df=cian_df.drop('Цена за квадрат', axis=1)

### Обработка признака "Адрес"

#Разобьём адрес на район, микрорайон, улицу. Это будет служить для дальнейшего заполнения пропусков:
cian_df['Район'] = cian_df['Адрес'].apply(lambda x: str(x.split(',')[1:2]).replace('[','').replace(']','').replace("'",''))
cian_df['Микрорайон'] = cian_df['Адрес'].apply(lambda x: str(x.split(',')[2:3]).replace('[','').replace(']','').replace("'",''))
cian_df['Улица'] = cian_df['Адрес'].apply(lambda x: str(x.split(',')[3:4]).replace('[','').replace(']','').replace("'",''))

### Обработка признака "Этаж"

#Создадим признак "этажей в доме":
cian_df['Этажей в доме'] = cian_df.Этаж.apply(lambda x: int(x.split(' ')[2]))

cian_df.Этаж = cian_df.Этаж.apply(lambda x: int(x.split(' ')[0]))

### Обработка признака "Время до метро"

#### **Переводим время до метро на транспорте в время до метро пешком**

#Для тех значений, где метро откроется в 2022 году, изначально поставим NaN
cian_df['Время до метро'][cian_df['Время до метро']==' откроется в 2022']=cian_df['Время до метро'][cian_df['Время до метро']==' откроется в 2022'].apply(lambda x: np.nan)

**Соотношение между временем на транспорте и временем пешком, на основе анализа по яндекс-картам (приблизительно):**

20 мин на транспорте - 20 мин пешком

30 тр - 45 пеш

45 тр - 90 пеш

90 тр - 280 пеш


#На основе этих данных написана функция по переводу минут на транспорте в минуты пешком:
def time_more_than_twenty(x):
    for i in range (1,11):    
        if x==20+i:            #если время на транспорте до метро занимает от 21 до 30 минут включительно
            x=x*(1.04**i)
    for k in range (1,16):
        if x==30+k:            #если время на транспорте до метро занимает от 31 до 45 минут включительно
            x=x*1.5*(1.02**k)
    for m in range (1,46):
        if x==45+m:            #если время на транспорте до метро занимает от 46 до 90 минут включительно
            x=x*2*(1.01**m)
    return x

def on_transport(x):
    if 'на транспорте' in x:
        return x

time_to_the_subway=cian_df['Время до метро'].dropna()

time_to_the_subway_on_transport=time_to_the_subway.apply(on_transport)

time_to_the_subway_on_transport = time_to_the_subway_on_transport.dropna()

time_to_the_subway_on_transport = time_to_the_subway_on_transport.apply(lambda x:int(x.replace(' мин. на транспорте','').replace(' ','').replace('⋅','').replace('<','')))

np.sort(time_to_the_subway_on_transport[time_to_the_subway_on_transport>20].unique())

time_to_the_subway_on_transport[time_to_the_subway_on_transport>20] = time_to_the_subway_on_transport[time_to_the_subway_on_transport>20].apply(time_more_than_twenty)

time_to_the_subway[time_to_the_subway_on_transport.index]=time_to_the_subway_on_transport

time_to_the_subway = time_to_the_subway.apply(lambda x:str(x))

time_to_the_subway = time_to_the_subway.apply(lambda x:float(x.replace(' мин. пешком','').replace(' ','').replace('⋅','').replace('<','')))

cian_df['Время до метро'][time_to_the_subway.index]=time_to_the_subway

### Обработка признака "Построен"

cian_df['Построен'].isna().sum()

cian_df['Год постройки'].isna().sum()

#Может быть так, что в одной из двух колонок присутствует значение года постройки. 
#Проверим это и обоюдно дополним данные:
cian_df['Построен'][cian_df['Год постройки'].dropna().index]=cian_df['Год постройки'][cian_df['Год постройки'].dropna().index]
cian_df['Год постройки'][cian_df['Построен'].dropna().index]=cian_df['Построен'][cian_df['Построен'].dropna().index]

#Уберём лишний признак:
cian_df = cian_df.drop('Год постройки',axis=1)

### Обработка признака "Тип жилья" и "Построен"

#Год постройки в новостройках отсутствует, но на циане есть информация, что год постройки - 2022. 
#Вобьём его для пропущенных значений "Тип жилья":
cian_df['Построен'][cian_df['Тип жилья'].isnull()] = cian_df['Построен'][cian_df['Тип жилья'].isnull()].fillna(2022.0)

#Если год постройки = 2022, соответственно тип жилья - новостройка. Заполним пропуски:
cian_df['Тип жилья'][(cian_df['Тип жилья'].isnull())&(cian_df['Построен']==2022.0)] = cian_df['Тип жилья'][(cian_df['Тип жилья'].isnull())&(cian_df['Построен']==2022.0)].fillna('Новостройка')

#Оставшиеся пропуски - вторички:
cian_df['Тип жилья'][cian_df['Тип жилья'].isnull()] = cian_df['Тип жилья'][cian_df['Тип жилья'].isnull()].fillna('Вторичка')

def rename1(x):
    if x == 'Вторичка Апартаменты':
        a='Вторичка'
        return a
    if x == 'Вторичка Пентхаус':
        a='Вторичка'
        return a
    if x == 'Новостройка Апартаменты':
        a='Новостройка'
        return a 
    if x == 'Новостройка Пентхаус':
        a='Новостройка'
        return a
    else:
        return x

#Переименуем типы жилья, чтобы на выходе получилось 2 типа: Новостройка и Вторичка:
cian_df['Тип жилья'] = cian_df['Тип жилья'].apply(rename1)

#Заполним пропуски года постройки для новостроек:
cian_df['Построен'][cian_df['Тип жилья']=='Новостройка']=cian_df['Построен'][cian_df['Тип жилья']=='Новостройка'].fillna(2022.0)

### Обработка признака "Высота потолков"

cian_df['Высота потолков'][cian_df['Высота потолков'].notnull()] = cian_df['Высота потолков'][cian_df['Высота потолков'].notnull()].apply(lambda x: float(x.replace(',','.').replace(' м','')))

cian_df['Высота потолков'].min()

cian_df['Высота потолков'].max()

#Заменим высоту потолков в данных ниже минимальной (2,4 м) по России на эту высоту:
cian_df['Высота потолков'][cian_df['Высота потолков']<2.4] = 2.4

#Поставим пропущенные значения для студий, 1,2,3-х комнатных квартир, где высота по данным выше 7 метров:
cian_df['Высота потолков'][(cian_df['Высота потолков']>7.0)&(cian_df['Название']!='Многокомнатная квартира')]=np.nan

### Обработка признака "Лифты"

#Заполним пропуски для пропущенных значений, где количество этажей в доме =5 и меньше
cian_df['Лифты'][(cian_df['Лифты'].isnull())&(cian_df['Этажей в доме']<6)] = cian_df['Лифты'][(cian_df['Лифты'].isnull())&(cian_df['Этажей в доме']<6)].fillna('Нет')

cian_df.head(3)

## ОБРАБОТКА ПРОПУСКОВ ПО АДРЕСУ, УЛИЦЕ, МИКРОРАЙОНУ, РАЙОНУ

cian_df.isna().sum()

#Список имён категориальных признаков:
list_categorical_features = ['Метро','Санузел','Балкон/лоджия','Тип дома','Тип перекрытий','Лифты','Отопление']

#Список имён числовых признаков:
list_numeric_features = ['Время до метро','Построен','Высота потолков','Подъезды']  

#Части адреса, которые по порядку будет проходить цикл:
parts_of_adress=['Адрес','Улица','Микрорайон','Район']                            

#Заполним пропущенные значения категориального признака самым частым значением в зависимости 
#от адреса, затем - от улицы, затем - от микрорайона, оставшиеся (если остались) - от района:

for feature in list_categorical_features:
    for part in parts_of_adress:
        try:
            for title in cian_df[part][cian_df[feature].isnull()].unique():
                try:
                    cian_df[feature][(cian_df[feature].isnull())&(cian_df[part]==title)]=cian_df[feature][cian_df[part]==title].value_counts().index[0]
                except:
                    pass
        except:
            pass   

#Заполним пропущенные значения числового признака средним значением в зависимости 
#от адреса, затем - от улицы, затем - от микрорайона, оставшиеся(если остались) - от района:

for feature in list_numeric_features:
    for part in parts_of_adress:
        try:
            for title in cian_df[part][cian_df[feature].isnull()].unique():
                try:
                    cian_df[feature][(cian_df[feature].isnull())&(cian_df[part]==title)]=cian_df[feature][cian_df[part]==title].mean()
                except:
                    pass
        except:
            pass

cian_df.isna().sum()

## Окончательная обработка

#Переведём в целые значения время до метро:
cian_df['Время до метро'] = cian_df['Время до метро'].apply(lambda x: int(x))

#Округлим до десятых значения высоты потолков:
cian_df['Высота потолков'] = cian_df['Высота потолков'].apply(lambda x: np.round(x,1))

#Округлим до целого значения количество подъездов:
cian_df['Подъезды'] = cian_df['Подъезды'].apply(lambda x: np.round(x,0))

cian_df['Лифты'].unique()

#Нас по большей части интересуют именно пассажирские лифты, поэтому если они есть, переименовываем в "Есть"
#Если нет или только грузовые - "Нет":
def lifts_rename(x):
    if x == 'Есть грузовой' or x=='2 груз.' or x=='3 груз.' or x=='4 груз.':
        a='Нет'
        return a
    if x == 'Есть':
        return x
    if x == 'Нет':
        return x
    else:
        a='Есть'
        return a

cian_df['Лифты'] = cian_df['Лифты'].apply(lifts_rename)

cian_df['Лифты'].value_counts()

### Обработка признака "Аварийность"

cian_df['Аварийность'].value_counts()

#1 дом из всех объявлений авариен. Признак бесполезен, удаляем:
cian_df = cian_df.drop('Аварийность',axis=1)

#Также признак не имеет полезных данных
cian_df = cian_df.drop('Площадь комнат+ обозначение смежных комнат- обозначение изолированных комнат',axis=1)

### Обработка признака "Вид из окон" и "Ремонт"

cian_df['Вид из окон'].value_counts()

cian_df['Ремонт'].value_counts()

#Так как ремонт сугубо индивидуален, а вид из окна весьма трудно спрогнозировать, удалим их:
cian_df = cian_df.drop(['Ремонт','Вид из окон'],axis=1)

cian_df.isna().sum()

cian_df.head()

cian_df.to_csv('cian_data_clean.csv', index=False)
